agents/iteration-controller/requirements.txt
# Iteration Controller MCP Server dependencies
# Install: pip install -r requirements.txt

# MCP SDK
mcp[cli]>=1.0.0

# Environment loading
python-dotenv>=1.0.0
